Fixes #237 - player history struct now populated from p_hist.txt. Since this is only used in the server, we don't need to revamp the client for it.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@321 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/common/md5.c
@@ -333,7 +333,7 @@ unsigned int len;
 extern void MD5Password (char *string)
 {
   MD5_CTX context;
-  unsigned int len = strlen (string);
+  unsigned int len = (unsigned int)strlen (string);
   char temp[MAX_PASS_LEN];
   int i;
   char* dst;
